# dict: Change to a sticky Copy-On-Write based table.