@trotterdylan 2017-02-20
This is really cool. Thanks for preparing this PR. It's a bit involved so I may take a few days to get through it.

At a high level, I wanted to point out that the immutability of dictEntry objects is an important invariant of the current design. I can't remember all the details off hand but I think it ultimately comes down to not being able to update both the key and value atomically. Can you comment on why this is no longer important in this design?