@nairb774 2017-02-21
Sorry about the size. It wouldn't surprise me to take some time - the nice part is that this is a performance change, not new functionality so time is on our side.
 
As an overview, the implementation is a blending of the table layout of the Python dict implementation, and a concurrency implementation borrowed from https://github.com/golang/sync/commit/54b13b0. In that light, writes to the table are serialized by a mutex, while reads are performed from a immutable table (assuming the map has transitioned to fast reads). This means that for heavy read-only dicts (think globals) most reads only have the cost of a single atomic read. For heavy read/write dicts, the cost mostly the overhead of locking the `recursiveMutex`.

If anything I write here would make better sense in the code or commit comment, let me know. I'd also be happy to add/modify existing benchmarks if you think that would be good to make sure this is as good an improvement as I think it is. The main goal for me has been to try to get the [CallSimple](https://github.com/google/grumpy/blob/004b7a8/benchmarks/bm_call_simple.py) benchmark closer to the speed that CPython is able to execute it (12ms).